Finished api call and fill array
