+++ app/src/main/java/com/example/randomuserapi/features/listofusers/listofusersactivity/ListOfUsersActivity.kt
@@ -1,11 +1,13 @@
 package com.example.randomuserapi.features.listofusers.listofusersactivity
 
+import android.content.Context
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.Toast
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView.LayoutManager
 import com.example.randomuserapi.calls.randomuserusecase.RandomUserUseCase
 import com.example.randomuserapi.calls.usecaseclasses.randomuserclass.RandomUser
 import com.example.randomuserapi.calls.usecaseclasses.randomuserentities.RandomUserEntity
@@ -21,6 +23,9 @@ class ListOfUsersActivity : AppCompatActivity(), ListOfUsersActivityInterface {
 
     lateinit var listOfUsersAdapter: ListOfUserAdapter
     lateinit var recyclerView: RecyclerView
+    lateinit var layoutManager: LayoutManager
+
+    lateinit var arrayOfUsers: ArrayList<RandomUser>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -33,15 +38,15 @@ class ListOfUsersActivity : AppCompatActivity(), ListOfUsersActivityInterface {
     fun initializeUI(){
         this.viewModel = ListOfUsersViewModel()
         this.viewModel.initializeViewModel()
-        this.viewModel.getUserList(5,applicationContext)
-        prepareUserList()
+        this.getUserList(50,applicationContext)
+        this.arrayOfUsers = ArrayList()
     }
 
     override fun prepareUserList() {
 
-        this.listOfUsersAdapter = ListOfUserAdapter(this.viewModel.getArrayOfUsersFromCall())
+        this.listOfUsersAdapter = ListOfUserAdapter(arrayOfUsers)
 
-        val layoutManager = LinearLayoutManager(applicationContext)
+        this.layoutManager = LinearLayoutManager(applicationContext)
 
         this.recyclerView = this.binding.rvRandomUserList
         this.recyclerView.layoutManager = layoutManager
@@ -58,6 +63,43 @@ class ListOfUsersActivity : AppCompatActivity(), ListOfUsersActivityInterface {
         }
 
          */
+    }
+
+    override fun getUserList(numberOfUsers: Int, context: Context) {
+
+        val fetchRandomUserData = Job()
+
+        val errorHandler = CoroutineExceptionHandler{ coroutineContext, throwable ->
+            Toast.makeText(context, "${throwable.message}: Ha ocurrido un error y no se pueden obtener más usuarios.", Toast.LENGTH_LONG).show()
+            println("Error ---> ${throwable.message}")
+            arrayOfUsers.add(RandomUser(
+                "Ms",
+                "Jhon",
+                "Doe",
+                "",
+                "",
+                "",
+                "male",
+                "jhondoe@falsemail.com",
+                "0000-000~${arrayOfUsers.size}"
+            ))
+        }
+
+        val scope = CoroutineScope(fetchRandomUserData + Dispatchers.Main)
+
+        scope.launch(errorHandler){
+            var randomUser = viewModel.randomUserUseCase.getRandomUsersFromCall()
+            if(randomUser != null){
+                arrayOfUsers.add(viewModel.fromEntityToUser(randomUser))
+                println("Data obtained from ${arrayOfUsers.size} users.")
+                if(arrayOfUsers.size < numberOfUsers){
+                    getUserList(numberOfUsers,context)
+                } else {
+                    prepareUserList()
+                }
+            }
+        }
+
     }
 
 }
