+++ app/src/main/java/com/example/randomuserapi/features/listofusers/listofusersactivity/ListOfUsersActivityInterface.kt
@@ -1,6 +1,10 @@
 package com.example.randomuserapi.features.listofusers.listofusersactivity
 
+import android.content.Context
+
 interface ListOfUsersActivityInterface {
     fun prepareUserList()
     fun endLottieAndShowUserList()
+    fun getUserList(numberOfUsers: Int, context: Context)
+
 }
