+++ app/src/main/java/com/example/randomuserapi/features/listofusers/listofusersviewmodel/ListOfUsersViewModel.kt
@@ -14,49 +14,10 @@ import java.util.Random
 
 class ListOfUsersViewModel: ListOfUsersViewModelInterface {
 
-    var logTag = this::class.java.toString()
-
-    private lateinit var randomUserUseCase: RandomUserUseCase
-    private lateinit var viewModelArrayOfUsers: ArrayList<RandomUser>
+    lateinit var randomUserUseCase: RandomUserUseCase
 
     override fun initializeViewModel(){
         this.randomUserUseCase = RandomUserUseCase()
-        this.viewModelArrayOfUsers = ArrayList()
-    }
-
-    override fun getUserList(numberOfUsers: Int, context: Context) {
-
-        for(calls in 1..numberOfUsers){
-
-            val fetchRandomUserData = Job()
-
-            val errorHandler = CoroutineExceptionHandler{ coroutineContext, throwable ->
-                Toast.makeText(context, "${throwable.message}: Ha ocurrido un error y no se pueden obtener más usuarios.", Toast.LENGTH_LONG).show()
-                println("Error ---> ${throwable.message}")
-                viewModelArrayOfUsers.add(RandomUser(
-                    "Ms",
-                    "Jhon",
-                    "Doe",
-                    "",
-                    "",
-                    "",
-                    "male",
-                    "jhondoe@falsemail.com",
-                    "0000-000~${viewModelArrayOfUsers.size}"
-                ))
-            }
-
-            val scope = CoroutineScope(fetchRandomUserData + Dispatchers.Main)
-
-                scope.launch(errorHandler){
-                    var randomUser = randomUserUseCase.getRandomUsersFromCall()
-                    if(randomUser != null){
-                        viewModelArrayOfUsers.add(fromEntityToUser(randomUser))
-                        println("Data obtained from ${viewModelArrayOfUsers.size} users.")
-                    }
-                }
-        }
-
     }
 
     override fun fromEntityToUser(userEntity: RandomUserEntity): RandomUser {
@@ -71,10 +32,6 @@ class ListOfUsersViewModel: ListOfUsersViewModelInterface {
             userEntity.randomUserResultsEntity?.get(0)?.email,
             userEntity.randomUserResultsEntity?.get(0)?.phone
         )
-    }
-
-    override fun getArrayOfUsersFromCall(): ArrayList<RandomUser>{
-        return this.viewModelArrayOfUsers
     }
 
 }
